remove systematic fetch_all
it was introduced in a time when we add a big depth in  `git clone` and hope
to find a common ancestor

now with recent git version and partial clones and stuff, it seems counter productive
to put this fetch_all: slower, .git takes 7Gb instead of 2 and in some occasions
it never finish.

If you really need it, you can put it in spec.yaml and it will continue to work.
Again this commit, just stop to silently add the fetch_all in all cases.