ak sparse: keep all dirs appart addons
for odoo (src), we need to keep all the dirs and exclude ./addons/
the previous implementation was doing a `ls`. so it was not working
if src was initialized with `git clone --filter:blob=null` because
ls was returning nothing.

This implementation asks git to list all the dirs expected in .

also run a linter on the file